Fixed missing domain name support for services.
The services routine that parses domain names is missing a `\-` in the regex that the container version has. This prevents services with domain names containing hyphens from being properly parsed. i.e. a service with `Host('service.foo.bar')` is parsed and registered correctly, but `Host('service.fo-o.bar')` fails and is returned as null. 

Hope this helps!